requirements-airflow.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
altair==4.2.2
annotated-types==0.6.0
anyio==4.2.0
apache-airflow==2.8.1
apache-airflow-providers-apache-beam==5.5.0
apache-airflow-providers-common-io==1.2.0
apache-airflow-providers-common-sql==1.10.0
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-google==10.13.1
apache-airflow-providers-http==4.8.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-sqlite==3.7.0
apache-beam==2.53.0
apispec==6.4.0
argcomplete==3.2.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.7.2
asttokens==3.0.1
async-lru==2.2.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.7.0
cachelib==0.9.0
cachetools==5.3.2
cattrs==23.2.3
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
cloudpickle==2.2.1
colorama==0.4.6
colorlog==4.8.0
comm==0.2.2
ConfigUpdater==3.2
connexion==2.14.2
coverage==7.13.4
crcmod==1.7
cron_descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.7
db-dtypes==1.2.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.4.2
docopt==0.6.2
docutils==0.20.1
email-validator==1.3.1
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.2.1
fastavro==1.9.3
fasteners==0.19
fastjsonschema==2.21.2
Flask==2.2.5
Flask-AppBuilder==4.3.10
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
gcloud-aio-auth==4.2.3
gcloud-aio-bigquery==7.0.0
gcloud-aio-storage==9.0.0
gcsfs==2023.12.2.post1
google-ads==22.1.0
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.39.0
google-cloud-appengine-logging==1.4.0
google-cloud-audit-log==0.2.5
google-cloud-automl==2.12.0
google-cloud-batch==0.17.7
google-cloud-bigquery==3.16.0
google-cloud-bigquery-datatransfer==3.13.0
google-cloud-bigquery-storage==2.24.0
google-cloud-bigtable==2.22.0
google-cloud-build==3.22.0
google-cloud-compute==1.15.0
google-cloud-container==2.37.0
google-cloud-core==2.4.1
google-cloud-datacatalog==3.17.2
google-cloud-dataflow-client==0.8.6
google-cloud-dataform==0.5.5
google-cloud-dataplex==1.11.0
google-cloud-dataproc==5.8.0
google-cloud-dataproc-metastore==1.14.0
google-cloud-dlp==3.14.0
google-cloud-kms==2.20.0
google-cloud-language==2.12.0
google-cloud-logging==3.9.0
google-cloud-memcache==1.8.0
google-cloud-monitoring==2.18.0
google-cloud-orchestration-airflow==1.10.0
google-cloud-os-login==2.13.0
google-cloud-pubsub==2.19.0
google-cloud-redis==2.14.0
google-cloud-resource-manager==1.11.0
google-cloud-run==0.10.1
google-cloud-secret-manager==2.17.0
google-cloud-spanner==3.41.0
google-cloud-speech==2.23.0
google-cloud-storage==2.14.0
google-cloud-storage-transfer==1.10.0
google-cloud-tasks==2.15.0
google-cloud-texttospeech==2.15.1
google-cloud-translate==3.14.0
google-cloud-videointelligence==2.12.0
google-cloud-vision==3.5.0
google-cloud-workflows==1.13.0
google-crc32c==1.5.0
google-re2==1.1
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
great-expectations==0.18.8
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-gcp==0.2.2
grpcio-status==1.60.0
gunicorn==21.2.0
h11==0.16.0
hdfs==2.7.3
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflection==0.5.1
iniconfig==2.3.0
ipykernel==6.29.4
ipython==8.38.0
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.3
Js2Py==0.74
json-merge-patch==0.2
json5==0.13.0
jsonpatch==1.33
jsonpickle==3.0.2
jsonpointer==3.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
lazy-object-proxy==1.10.0
limits==3.7.0
linkify-it-py==2.0.2
lockfile==0.12.2
looker-sdk==23.20.1
makefun==1.16.0
Mako==1.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.2.0
multidict==6.0.4
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.4
notebook_shim==0.2.4
numpy==1.24.4
oauthlib==3.2.2
objsize==0.6.1
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
ordered-set==4.1.0
orjson==3.9.10
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandas-gbq==0.20.0
pandocfilters==1.5.1
parso==0.8.6
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
pip==26.0.1
platformdirs==4.9.2
pluggy==1.6.0
prison==0.2.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.7
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydata-google-auth==1.8.2
pydot==1.4.2
Pygments==2.17.2
pyjsparser==2.7.1
PyJWT==2.8.0
pymongo==4.6.1
pyOpenSSL==23.3.0
pyparsing==3.1.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-mock==3.15.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-json-logger==4.0.0
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.0
rich-argparse==1.4.0
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.17.17
scipy==1.15.3
Send2Trash==2.1.0
setproctitle==1.3.3
setuptools==66.1.1
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.8.3
SQLAlchemy==1.4.51
sqlalchemy-bigquery==1.9.0
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.6.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
time-machine==2.13.0
tinycss2==1.4.0
tomli==2.4.0
toolz==1.1.0
tornado==6.4.1
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uc-micro-py==1.0.2
unicodecsv==0.14.1
universal-pathlib==0.1.4
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.7
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==2.2.3
wheel==0.46.3
widgetsnbextension==4.0.15
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.17.0
zstandard==0.22.0
